requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
Django==4.2.21
djangorestframework==3.16.0
frozenlist==1.6.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
multidict==6.4.4
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
zope.interface==7.2
